# compare buffer bytes directly for faster detection