#### compare buffer bytes directly for faster detection

Previously, in order to detect the type of the file, the buffer was
converted to ASCII strings and then compared to hard-coded strings. Now,
the bytes of the buffer are compared directly to hard-coded bytes, which
is more efficient.

    name                        old time/op  new time/op  delta
    Calipers/png-304x85         90.9µs ± 3%  85.6µs ± 3%   -5.85%  (p=0.008 n=5+5)
    Calipers/png-304x85-buffer  2.14µs ± 8%  1.29µs ± 3%  -39.67%  (p=0.008 n=5+5)

#### reduce number of benchmark cases

Previously the benchmark script ran a benchmark for each PNG file in
`test/fixtures`. The measurements for all PNGs were roughly the same, so
running a benchmark for each seems redundant. It was also time
consuming. Now, only a single PNG is benchmarked.